+++ packages/preview-middleware-client/test/unit/adp/controllers/AddFragment.controller.test.ts
@@ -102,6 +102,80 @@ describe('AddFragment', () => {
 
             expect(openSpy).toHaveBeenCalledTimes(1);
         });
+
+        test('fills json model with data and defaultAggregationArrayIndex', async () => {
+            const overlays = {
+                getId: jest.fn().mockReturnValue('some-id')
+            };
+
+            ControlUtils.getRuntimeControl = jest.fn().mockReturnValue({
+                getMetadata: jest.fn().mockReturnValue({
+                    getAllAggregations: jest.fn().mockReturnValue({
+                        'tooltip': {},
+                        'customData': {},
+                        'layoutData': {},
+                        'dependents': {},
+                        'dragDropConfig': {},
+                        'content': {}
+                    }),
+                    getDefaultAggregationName: jest.fn().mockReturnValue('content'),
+                    getName: jest.fn().mockReturnValue('Toolbar')
+                })
+            });
+
+            ControlUtils.getControlAggregationByName = jest
+                .fn()
+                .mockReturnValue({ 0: {}, 1: {}, 2: {}, 3: {}, 4: {}, 5: {}, 6: {} });
+
+            const overlayControl = {
+                getDesignTimeMetadata: jest.fn().mockReturnValue({
+                    getData: jest.fn().mockReturnValue({
+                        aggregations: { content: { actions: { move: null }, domRef: ':sap-domref' } }
+                    })
+                })
+            };
+            sapCoreMock.byId.mockReturnValue(overlayControl);
+
+            OverlayRegistry.getOverlay = jest.fn().mockReturnValue({
+                getDesignTimeMetadata: jest.fn().mockReturnValue({
+                    getData: jest.fn().mockReturnValue({
+                        aggregations: {}
+                    })
+                })
+            });
+
+            const addFragment = new AddFragment(
+                'adp.extension.controllers.AddFragment',
+                overlays as unknown as UI5Element,
+                {} as unknown as RuntimeAuthoring,
+                {
+                    title: 'ADP_ADD_FRAGMENT_DIALOG_TITLE',
+                    defaultAggregationArrayIndex: 1
+                }
+            );
+
+            const openSpy = jest.fn();
+            const setPropertySpy = jest.fn();
+            addFragment.model = {
+                setProperty: setPropertySpy
+            } as unknown as JSONModel;
+            await addFragment.setup({
+                setEscapeHandler: jest.fn(),
+                destroy: jest.fn(),
+                setModel: jest.fn(),
+                open: openSpy,
+                close: jest.fn()
+            } as unknown as Dialog);
+
+            const escapeHandlerCb = (addFragment.dialog.setEscapeHandler as jest.Mock).mock.calls[0][0];
+
+            escapeHandlerCb({ resolve: jest.fn() });
+
+            expect(openSpy).toHaveBeenCalledTimes(1);
+            const lastCall = setPropertySpy.mock.calls[setPropertySpy.mock.calls.length - 1];
+            expect(lastCall[0]).toBe('/selectedIndex');
+            expect(lastCall[1]).toBe(1);
+        });
     });
 
     describe('onAggregationChanged', () => {